requirements.txt
pytorch-lightning
torch
huggingface
lightning
accelerate
bitsandbytes
datasets
peft
trl
scikit-learn
flash-attn
tensorboard
tensorboardX
sentence_transformers
tiktoken
blobfile
sentencepiece
bitsandbytes
fasttext
easynmt